requirements.txt
easydict>=1.9
matplotlib>=3.7.1
networkx>=3.1
numpy>=1.25.2
pandas>=1.5.3
PyYAML>=6.0.2
scipy>=1.13.0
torch>=2.0.1
torch_geometric>=2.6.1
tqdm>=4.65.0
